API/Daemon: Unix Socket/Windows TCP.
Split up the socket and transport creation depending on the OS. For
Windows we'll use TCP, for anything else we'll use Unix Domain Sockets.
