+++ daemon/socket/proxy.go
@@ -9,8 +9,6 @@ import (
 	"net/http"
 	"net/http/httputil"
 	"net/url"
-	"os"
-	"path/filepath"
 	"strings"
 	"time"
 
@@ -149,39 +147,6 @@ func requestIDHandler(next http.Handler) http.Handler {
 		r = r.WithContext(ctx)
 		next.ServeHTTP(w, r)
 	})
-}
-
-func makeSocket(socketPath string, groupShared bool) (net.Listener, error) {
-	absPath, err := filepath.Abs(socketPath)
-	if err != nil {
-		return nil, err
-	}
-
-	// Attempt to remove an existing socket at this path if it exists.
-	// Guarding against a server already running is outside the scope of this
-	// module.
-	err = os.Remove(absPath)
-	if err != nil && !os.IsNotExist(err) {
-		return nil, err
-	}
-
-	l, err := net.Listen("unix", absPath)
-	if err != nil {
-		return nil, err
-	}
-
-	mode := os.FileMode(0700)
-	if groupShared {
-		mode = 0760
-	}
-
-	// Does not guarantee security; BSD ignores file permissions for sockets
-	// see https://github.com/manifoldco/torus-cli/issues/76 for details
-	if err = os.Chmod(socketPath, mode); err != nil {
-		return nil, err
-	}
-
-	return l, nil
 }
 
 // proxyCanceler supports canceling proxied requests via a timeout, and
