+++ daemon/socket/socket_windows.go
@@ -1,0 +1,9 @@
+// +build windows
+
+package socket
+
+import "net"
+
+func makeSocket(socketPath string, groupShared bool) (net.Listener, error) {
+	return net.Listen("tcp", ":50")
+}
